feat: Add query and fragment support to HttpRequest class
The code changes in `HttpRequest.hpp` and `HttpRequest.cpp` files introduce support for query and fragment in the `HttpRequest` class. The `parseMethodLine` function now parses the query and fragment from the request line, and the `print` function displays the query and fragment in the output. This enhancement aims to improve the handling of HTTP requests by capturing additional information from the request URI.

currently up to standard RFC1945
new uri is RFC3986 but we only have to comply with the rfc for http1.0/1.1